+++ src/main/kotlin/com/github/ilovegamecoding/intellijcodexp/activities/CodeXPStartupActivity.kt
@@ -1,14 +1,14 @@
 package com.github.ilovegamecoding.intellijcodexp.activities
 
+import com.github.ilovegamecoding.intellijcodexp.managers.CodeXPUIManager
 import com.github.ilovegamecoding.intellijcodexp.services.CodeXPService
 import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
 
 class CodeXPStartupActivity : StartupActivity {
     override fun runActivity(project: Project) {
-        thisLogger().warn("CodeXPStartupActivity.runActivity")
-        val codeXPService = ApplicationManager.getApplication().getService(CodeXPService::class.java)
+        ApplicationManager.getApplication().getService(CodeXPService::class.java)
+        CodeXPUIManager.createDialogArea()
     }
 }